pyproject.toml
[project]
name = "weather-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.227.0",
    "constructs>=10.4.3",
    "dlt[athena,duckdb,workspace]>=1.18.2",
    "notebook>=7.5.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
